+++ RandomSkunk.Results/Result{T}.cs
@@ -297,7 +297,7 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return new CompositeError(
+                return new AggregateError(
                     [GetError(), error],
                     $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
@@ -308,7 +308,7 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return new CompositeError(
+                return new AggregateError(
                     [GetError(), error],
                     $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
@@ -344,7 +344,7 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return new CompositeError(
+                return new AggregateError(
                     [GetError(), error],
                     $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
@@ -355,7 +355,7 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return new CompositeError(
+                return new AggregateError(
                     [GetError(), error],
                     $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
@@ -566,13 +566,13 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
@@ -606,13 +606,13 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
@@ -647,13 +647,13 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
@@ -688,13 +688,13 @@ public readonly struct Result<T> : IEquatable<Result<T>>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return new CompositeError(
+                    return new AggregateError(
                         [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }