+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning].
     - Change `Maybe<T>.None` from property to method.
     - Don't create nested error in `Error.FromException` method.
     - Rearrange `errorCode` and `identifier` parameters in `Error.FromException`, `Result.Fail`, and `Result<T>.Fail` methods.
-    - Replace `CompositeError.Create` factory method with public constructor.
-    - Rename `CompositeError.Errors` property to `InnerErrors`.
+    - Rename `CompositeError` to `AggregateError`.
+    - Replace `AggregateError.Create` factory method with public constructor.
+    - Rename `AggregateError.Errors` property to `InnerErrors`.
 
 - RandomSkunk.Results.AspNetCore:
     - In the `ToActionResult` extension methods, change the optional parameter from `Func<int, int>? getHttpStatusCode` to `Func<Error, IActionResult>? onFail`.